requirements.txt
altgraph==0.17.3
asgiref==3.5.2
autopep8==1.7.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colour==0.1.5
distlib==0.3.6
Django==4.1.2
django-shopping-cart==0.1
et-xmlfile==1.1.0
filelock==3.8.0
idna==3.4
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
macholib==1.16.2
MarkupSafe==2.1.1
numpy==1.23.3
openai==0.23.0
openpyxl==3.0.10
pandas==1.5.0
pandas-stubs==1.5.0.220926
Pillow==9.2.0
pipenv==2022.10.12
platformdirs==2.5.2
pycodestyle==2.9.1
pyinstaller==5.4.1
pyinstaller-hooks-contrib==2022.10
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tk==0.1.0
tkmacosx==1.0.5
toml==0.10.2
tqdm==4.64.1
types-pytz==2022.2.1.0
typing_extensions==4.3.0
urllib3==1.26.12
virtualenv==20.16.5
virtualenv-clone==0.5.7
Werkzeug==2.2.2
